Restructure repository and add markdown export
