+++ data/release_metrics_1.csv
@@ -1,0 +1,13 @@
+release,sentences,semantic_spread,redundancy_index,cluster_entropy
+Rel-8,565708,0.8358441632826105,0.829537883400917,5.9843244031576335
+Rel-9,586563,0.8351972925765683,0.8352721482515335,5.982103523958227
+Rel-10,608402,0.8318131099360998,0.8331997394561768,5.943574977662733
+Rel-11,687763,0.8316013819244352,0.825080931186676,5.965783027319289
+Rel-12,758512,0.8351794847249725,0.8347167372703552,6.0171571788445455
+Rel-13,903233,0.8308291366489271,0.8323939591646194,5.970197847962651
+Rel-14,863163,0.8343180422454071,0.8379383087158203,5.996686863121215
+Rel-15,1126870,0.8296580292954951,0.8340643048286438,6.01258856506894
+Rel-16,1243134,0.8320429399442864,0.8339799642562866,5.9424490802802135
+Rel-17,1517182,0.8287739629819761,0.8238424062728882,5.958959128860738
+Rel-18,997724,0.8200660185466644,0.8225340694189072,5.9454488556303415
+Rel-19,578786,0.8225662327026435,0.8200987428426743,5.9982232292976265
