--- delta_metrics.csv
@@ -1,12 +1,0 @@
-from,to,change_magnitude,novelty_density
-Rel-8,Rel-9,0.00035881996154785156,0.033503178507089615
-Rel-9,Rel-10,0.0004621148109436035,0.03151042014360428
-Rel-10,Rel-11,0.00030153989791870117,0.037604350596666336
-Rel-11,Rel-12,0.0005077719688415527,0.047221940010786057
-Rel-12,Rel-13,0.0012865662574768066,0.06050318852066994
-Rel-13,Rel-14,0.0011696815490722656,0.04028789699077606
-Rel-14,Rel-15,0.0016641020774841309,0.07824357599020004
-Rel-15,Rel-16,0.0011259913444519043,0.05075536295771599
-Rel-16,Rel-17,0.001286327838897705,0.06289168447256088
-Rel-17,Rel-18,0.007163107395172119,0.06366242468357086
-Rel-18,Rel-19(dev.),0.024420499801635742,0.11580195277929306
