--- delta_metrics_1.csv
@@ -1,12 +1,0 @@
-from,to,change_magnitude,novelty_density
-Rel-8,Rel-9,0.00035881996154785156,0.4071333110332489
-Rel-9,Rel-10,0.0004621148109436035,0.40635931491851807
-Rel-10,Rel-11,0.00030153989791870117,0.40756744146347046
-Rel-11,Rel-12,0.0005077719688415527,0.407968133687973
-Rel-12,Rel-13,0.0012865662574768066,0.41221991181373596
-Rel-13,Rel-14,0.001169741153717041,0.41197043657302856
-Rel-14,Rel-15,0.0016641020774841309,0.4132034182548523
-Rel-15,Rel-16,0.001125931739807129,0.4116007685661316
-Rel-16,Rel-17,0.001286327838897705,0.4106443226337433
-Rel-17,Rel-18,0.007163107395172119,0.38326117396354675
-Rel-18,Rel-19,0.024420499801635742,0.3957866132259369
