--- release_metrics.csv
@@ -1,13 +1,0 @@
-release,sentences,semantic_spread,redundancy_index,cluster_entropy
-Rel-8,565708,0.8358441631532293,0.8430780917406082,9.395446490695534
-Rel-9,586563,0.8351972924560565,0.8321858942508698,9.40388006085623
-Rel-10,608402,0.8318131098811887,0.827052503824234,9.459033282886741
-Rel-11,687763,0.8316013818680363,0.8296196609735489,9.537719139702713
-Rel-12,758512,0.8351794846557787,0.8331300169229507,9.608147163560163
-Rel-13,903233,0.8308291366249476,0.831301748752594,9.723303884067562
-Rel-14,863163,0.8343180422110517,0.832138791680336,9.699202723412943
-Rel-15,1126870,0.8296580292513022,0.8294869512319565,9.889410905948179
-Rel-16,1243134,0.8320429399172182,0.8286896049976349,9.954627400731155
-Rel-17,1517182,0.8287739629594386,0.8283903449773788,10.08770088477401
-Rel-18,997724,0.8200660184430315,0.8193341344594955,9.76649650180245
-Rel-19(dev.),578786,0.8225662327463037,0.8208694308996201,9.376456801753168
