--- requirement.txt
@@ -1,12 +1,0 @@
-numpy
-pandas
-torch
-spacy
-tqdm
-sentence-transformers
-scikit-learn
-scipy
-psutil
-pdfminer.six
-pynvml
-wandb
